refactor: centralize CLI login options
